Added the first actual Ritual and added some checks to prevent self-obliteration

+++ modular_blackmoor/zizo/_zizo_ritual.dm
@@ -53,8 +53,9 @@
 	if (consumption)
 		for (var/dir in found_reagents)
 			var/atom/movable/A = found_reagents[dir]
-			do_sparks(2, FALSE, A)
-			qdel(A)
+			if(!istype(A, /mob/living))//If the rite targets a human directly on the sigil it shouldn't consume them.
+				do_sparks(2, FALSE, A)
+				qdel(A)
 
 	//Returns the list of reagents found so that the actual ritual proc can use them
 	return found_reagents
